add base of behavior and minor refine in everywhere

+++ parsian_ai/include/parsian_ai/plays/playoff.h
@@ -447,7 +447,7 @@ private:
     bool isPlanDone();
     bool isPlanFaild();
     bool setTimer;
-    long tempStart;
+    unsigned int tempStart;
     ////////////////////////////
 
     bool isKickDone    (CRolePlayOff*);
